requirements.txt
# Firebird database driver
fdb>=2.0.0

# HTTP requests for cloud sync
requests>=2.28.0

# Environment variable loading from .env files
python-dotenv>=1.0.0
